gallery search bar and titles working
